% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPopulationNew.R
\name{GetPopulationNew}
\alias{GetPopulationNew}
\title{GetPopulationNew function
This function read population data from the EURO dashboard}
\usage{
GetPopulationNew()
}
\description{
GetPopulationNew function
This function read population data from the EURO dashboard
}
